requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
certifi==2022.5.18.1
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.4
django-allauth==0.50.0
django-cors-headers==3.12.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
dnspython==2.2.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
nltk==3.7
numpy==1.22.4
oauthlib==3.2.0
pandas==1.4.2
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pymongo==3.12.1
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.2.4
textblob==0.17.1
tqdm==4.64.0
tweepy==4.10.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing-extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.1.0
zope.interface==5.4.0
